# packages/shiori-api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "shiori",
    "description": "TODO DESCRIPTION OF SHIORI",
    "license": {
      "name": ""
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/libraries": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "List all libraries.",
        "operationId": "list_libraries",
        "responses": {
          "200": {
            "description": "Successfully fetched libraries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "name",
                      "path",
                      "created_at"
                    ],
                    "properties": {
                      "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Timestamp of when the media was created.",
                        "examples": [
                          "2024-11-08T17:23:41Z"
                        ]
                      },
                      "id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Unique identifier for the library.",
                        "examples": [
                          86
                        ]
                      },
                      "name": {
                        "type": "string",
                        "description": "Name of the library.",
                        "examples": [
                          "Light Novels"
                        ]
                      },
                      "path": {
                        "type": "string",
                        "description": "File system path to the library's directory.",
                        "examples": [
                          "/data/books/light_novels"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "library"
        ],
        "summary": "Create a new library.",
        "description": "The directory must already exist on the system.",
        "operationId": "create_library",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "path"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the library.",
                    "examples": [
                      "Light Novels"
                    ]
                  },
                  "path": {
                    "type": "string",
                    "description": "File system path to the library's directory, relative to the application's base directory.",
                    "examples": [
                      "light_novels"
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully created library",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "path",
                    "created_at"
                  ],
                  "properties": {
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Timestamp of when the media was created.",
                      "examples": [
                        "2024-11-08T17:23:41Z"
                      ]
                    },
                    "id": {
                      "type": "integer",
                      "format": "int32",
                      "description": "Unique identifier for the library.",
                      "examples": [
                        86
                      ]
                    },
                    "name": {
                      "type": "string",
                      "description": "Name of the library.",
                      "examples": [
                        "Light Novels"
                      ]
                    },
                    "path": {
                      "type": "string",
                      "description": "File system path to the library's directory.",
                      "examples": [
                        "/data/books/light_novels"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/libraries/{id}/media": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "List library media.",
        "operationId": "list_library_media",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the library",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched library media",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "name",
                      "size",
                      "path",
                      "extension",
                      "created_at",
                      "library_id"
                    ],
                    "properties": {
                      "cover_path": {
                        "type": [
                          "string",
                          "null"
                        ],
                        "description": "Endpoint where the cover is stored.",
                        "examples": [
                          "/api/v1/media/4/cover"
                        ]
                      },
                      "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Timestamp of when the media was created.",
                        "examples": [
                          "2026-03-23T12:00:00Z"
                        ]
                      },
                      "extension": {
                        "type": "string",
                        "description": "File extension of the media.",
                        "examples": [
                          "epub"
                        ]
                      },
                      "id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Unique identifier for the media item.",
                        "examples": [
                          86
                        ]
                      },
                      "library_id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Library this media belongs to.",
                        "examples": [
                          2
                        ]
                      },
                      "name": {
                        "type": "string",
                        "description": "Name of the media file, excluding extension.",
                        "examples": [
                          "86_Volume_1"
                        ]
                      },
                      "path": {
                        "type": "string",
                        "description": "File system path where the media is stored.",
                        "examples": [
                          "/data/books/light_novels/86_Volume_1.epub"
                        ]
                      },
                      "size": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Size of the media file in bytes.",
                        "examples": [
                          "102400"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Library not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "library"
        ],
        "summary": "Upload a new media item to the specified library.",
        "operationId": "create_library_media",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the library",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "files"
                ],
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "binary",
                      "minimum": 0
                    },
                    "description": "An array of files to upload."
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully added media to the library",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "name",
                      "size",
                      "path",
                      "extension",
                      "created_at",
                      "library_id"
                    ],
                    "properties": {
                      "cover_path": {
                        "type": [
                          "string",
                          "null"
                        ],
                        "description": "Endpoint where the cover is stored.",
                        "examples": [
                          "/api/v1/media/4/cover"
                        ]
                      },
                      "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Timestamp of when the media was created.",
                        "examples": [
                          "2026-03-23T12:00:00Z"
                        ]
                      },
                      "extension": {
                        "type": "string",
                        "description": "File extension of the media.",
                        "examples": [
                          "epub"
                        ]
                      },
                      "id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Unique identifier for the media item.",
                        "examples": [
                          86
                        ]
                      },
                      "library_id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Library this media belongs to.",
                        "examples": [
                          2
                        ]
                      },
                      "name": {
                        "type": "string",
                        "description": "Name of the media file, excluding extension.",
                        "examples": [
                          "86_Volume_1"
                        ]
                      },
                      "path": {
                        "type": "string",
                        "description": "File system path where the media is stored.",
                        "examples": [
                          "/data/books/light_novels/86_Volume_1.epub"
                        ]
                      },
                      "size": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Size of the media file in bytes.",
                        "examples": [
                          "102400"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid uploaded media"
          },
          "404": {
            "description": "Library not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}